Correct profiling result
The previous profiling result is not correct because I foget to
synchronize the CUDA threads before counting the time. Also, I should
use time measurement instead of clock measurement because the CPU might
be deschedueld by the OS whilte it is waiting for the GPU. In this case,
we would prefer knowing the actual time eplapsed. I also captured
another sequence which is already uploaded in Box.